+++ helpers/process-dita.js
@@ -3,26 +3,27 @@ const fs = require('fs');
 const path = require('path');
 const { zipSync } = require('cross-zip');
 
-function installDITAPlugin(pluginPath, silent) {
+function installDITAPlugin(ditaExecPath, pluginPath, silent) {
     console.log(`> Installing plugin ${pluginPath}`);
 
     const pluginName = path.basename(pluginPath);
     try {
         if (!fs.existsSync(pluginPath)) {
-            shell.exec(`dita install ${pluginName}`, { silent });
+            console.log(`${ditaExecPath} install ${pluginName}`);
+            shell.exec(`${ditaExecPath} install ${pluginName}`, { silent });
         } else if (fs.statSync(pluginPath).isFile()) {
             // ZIP file
-            shell.exec(`dita install ${pluginPath}`, { silent });
+            shell.exec(`${ditaExecPath} install ${pluginPath}`, { silent });
         } else if (fs.statSync(pluginPath).isDirectory()) {
             console.info(
                 `> Detected local plugin ${pluginName}. Reinstalling...`
             );
-            shell.exec(`dita uninstall ${pluginName}`, {
+            shell.exec(`${ditaExecPath} uninstall ${pluginName}`, {
                 silent,
             });
             const zipPath = path.join(shell.tempdir(), pluginName + '.zip');
             zipSync(pluginPath, zipPath);
-            shell.exec(`dita install ${zipPath}`, {
+            shell.exec(`${ditaExecPath} install ${zipPath}`, {
                 silent,
             });
         }
@@ -33,15 +34,22 @@ function installDITAPlugin(pluginPath, silent) {
     }
 }
 
-module.exports = function (configPath, config, silent = true) {
+module.exports = function (ditaExecPath, configPath, config, silent = true) {
     const configDir = path.dirname(path.resolve(process.cwd(), configPath));
     if (config.plugins && config.plugins.length) {
         for (const pluginPath of config.plugins)
-            installDITAPlugin(path.join(configDir, pluginPath), silent);
+            installDITAPlugin(
+                ditaExecPath,
+                path.join(configDir, pluginPath),
+                silent
+            );
     }
 
     shell.exec(
-        `dita -f ${config.transtype} -i ${path.join(configDir, config.input)}`,
+        `${ditaExecPath} -f ${config.transtype} -i ${path.join(
+            configDir,
+            config.input
+        )}`,
         {
             silent,
             fatal: true,